implemented GraphSearchAStar.  fixed IHeuristic and finished HeuristicEuclid.

+++ Burton.Lib.Graph/IHeuristic.cs
@@ -6,6 +6,6 @@ namespace Burton.Lib.Graph
 {
     public interface IHeuristic<T>
     {
-        double Calculate(T Graph, int Node1, int Node2);
+        float Calculate(T Graph, int Node1, int Node2);
     }
 }
